Revert "chess board printing" aa
This reverts commit 126dfdf32584ae3162dac5fade0129b95d00d31e.

+++ Userland/SampleCodeModule/include/libc.h
@@ -45,6 +45,6 @@ int abs(int n);
 int strcmp(const char *str1, const char *str2);
 //IN: Numero a convertir, buffer a guardar el numero, cuantos bytes ocupa
 //OUT: Referencia al buffer con el numero guardado
-char *intToHexa(long long num, char *str, int bytes);
+char *intToHexa(long int num, char *str, int bytes);
 
 #endif
